Image ScrollView on HomePage ViewController
Added the logo image UIImageView object, and inserted it into the UIScrollView and made it dependent on the properties of the image so that it is scrolling left, right, up, down, and pinching the image. Still struggling with horizontal and vertical scaling, as well as testing inside of Xcode the actual pinching functionality, because the `alt`-key configs are outside of the bounds of the scroll view.